pyproject.toml
[project]
name = "genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "arxiv>=3.0.0",
    "chromadb>=1.5.7",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "jq>=1.11.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.28",
    "langchain-experimental>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "langgraph-cli[inmem]>=0.4.24",
    "matplotlib>=3.10.8",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.10.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.4.0",
    "tiktoken>=0.12.0",
    "unstructured[docx]>=0.22.21",
    "wikipedia>=1.4.0",
]
